requirements.txt
blinker==1.5
cachelib==0.9.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.4.0
googletrans==4.0.0rc1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
PyJWT==2.6.0
pymongo==4.3.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
stripe==5.0.0
twilio==7.15.2
urllib3==1.26.12
Werkzeug==2.2.2
